indexer/stream-to-queue-publisher/requirements.txt
python-dotenv~=1.0
websocket-client~=1.8
pika~=1.3
numpy~=1.26